Weather EDA Using Python
For this project, a parent dataset comprising of 4 years of electrical consumption, generation, pricing, and weather data for Spain has been used. The dataset used here has mainly the weather data, purchased as part of a personal project from the Open Weather API for the 5 largest cities in Spain. We will try to see the relationship between certain variables, the correlation (if any) between features, and generate visualizations to better comprehend our data at hand. This will help us run a post-modeling guardrail and check if our model prediction is in line with the Exploratory Data Analysis (EDA) visualizations.