EDA of Poverty Probability of Individuals
The goal of this project is to predict the probability status of poverty of individuals across seven different countries. The poverty line is set at a threshold of $2.50/day for a given set of social and economic indicators. The data is on the Poverty Probability Index, which predicts the probability of an individual to be under the poverty line using the characteristic of the household, assets they hold, and other financial indicators from the survey conducted by InterMedia.